fix(pricing): address review findings on PR #166
- Use Decimal throughout _generate_price_recommendations: target_margin,
  margin_pct, margin_gap, avg_gap, and total_impact now computed with
  Decimal arithmetic instead of float (CLAUDE.md rule #9)
- Restore SensitivityCalcResponse fields from float back to Decimal
- Revert deploy health-check to http://localhost:8000/health — the Caddy
  proxy URL added an unnecessary external dependency that could block
  valid deployments; the backend port is still exposed on the host
- Lazy-load products in pricing-page via effect() — products are now
  fetched only when the Demand & Mix tab is first opened, saving one
  API round-trip on every pricing page visit

+++ backend/src/pricing/schemas.py
@@ -194,15 +194,15 @@ class SensitivityCalcRequest(BaseModel):
 
 
 class SensitivityCalcResponse(BaseModel):
-    unit_cost_usd: float
-    fx_rate: float
-    landed_cost_ngn: float
-    selling_price: float
-    margin_pct: float
+    unit_cost_usd: Decimal
+    fx_rate: Decimal
+    landed_cost_ngn: Decimal
+    selling_price: Decimal
+    margin_pct: Decimal
     quantity: int
-    total_revenue: float
-    total_cost: float
-    gross_profit: float
+    total_revenue: Decimal
+    total_cost: Decimal
+    gross_profit: Decimal
 
 
 # ---------------------------------------------------------------------------
